dm-table: re-arrange inlinecrypt capability check logic
According to upstream commit 75c5d31, "a non-zero return code
returned by iterate_devices_callout_fn will stop the iteration in
advance". In addition, commit a36aead specifically inverts the
dax capability check to be is(not capable) instead of not(capable)
to address the issue. Therefore, follow that change to invert the
logic on inlinecrypt capability check as well.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: Ia6ca5aed174e781017299d626676e30f787d38a6
Signed-off-by: negrroo <mohammedaelnaggar1@gmail.com>